Changed project settings to support new hosting platform.

+++ api/classes/postgres_heroku_db.php
@@ -2,10 +2,10 @@
     class postgres_heroku_DBI{
         public $table_prefix = 'main';
 
-        protected $dbname= "d31a3bbhmd66q4";
-        protected $dbuser = "fmowufsnbrohyn";
-        protected $dbpassword = "918eba83c85790418ad9af9915e7270f403fb3df6cf26556da15e762ad1ca1d1";
-        protected $dbhost = "ec2-54-211-159-145.compute-1.amazonaws.com";
+        protected $dbname= "webapp-sandbox-server";
+        protected $dbuser = "server";
+        protected $dbpassword = "z0n1ED4y0s1qL8h9MVslSSXt0Nf";
+        protected $dbhost = "localhost";
         protected $dbport = "5432";
 
         protected $connectionstr;
@@ -34,6 +34,11 @@
             $this->_connect();
             $resource = pg_query($this->dbconn, $query);
             $results = pg_fetch_all($resource);
+            //pg_fetch_all returns an empty array when there are no results in php 8
+            //but false in php 7. this will make everything work for either case
+            if($results === false && pg_result_error($results) === false){
+                $results = array();
+            }
             $this->_disconnect();
             return $results;
         }
@@ -44,14 +49,13 @@
             $results = $this->query('SELECT * FROM information_schema.schemata');
             foreach($results as $schema) {
                 $schema_name = $schema['schema_name'];
-                if($schema_name != 'pg_catalog' && $schema_name != 'information_schema'){
+                if($schema_name != 'pg_catalog' && $schema_name != 'information_schema' && $schema_name != 'pg_toast'){
                     $dropsql = "DROP SCHEMA IF EXISTS $schema_name CASCADE";
                     if($this->query($dropsql) === false){
                         return false;
                     };
                 }
             }
-    
             //create new schema
             return($this->query("CREATE SCHEMA ".$this->table_prefix) !== false);
         }
